requirements/_base.txt
Django==1.5
djangorestframework==2.3.6
django-filter==0.6
Markdown==2.3.1
